# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "description": "API для управления аутентификацией пользователей",
        "contact": {
            "email": "gamletel@list.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список всех постов",
                "description": "Возвращает список всех постов",
                "operationId": "5f292ed47bdbfa79356750be0807450c",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создать новый пост",
                "description": "Создаёт новый пост и возвращает его данные",
                "operationId": "c2e0b0e26215e6d34ece33ebae25f7df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "text",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Название поста"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Краткое описание"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Содержимое поста"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить пост по ID",
                "description": "Возвращает данные поста по его ID",
                "operationId": "c8bd68bc05422c879bc2dd429ae5782b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Удалить пост",
                "description": "Удаляет пост по его ID",
                "operationId": "0f85c7086f379490b8a4c03db9171f76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пост удалён"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Posts"
                ],
                "summary": "Обновить пост",
                "description": "Обновляет данные поста по его ID",
                "operationId": "a41813e460c73b07ab4dd8cc19f80eac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "text"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновлённое название"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Обновлённое краткое описание"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Обновлённое содержимое"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "updated_image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{id}/comments": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить комментарии к посту",
                "description": "Возвращает список комментариев для указанного поста",
                "operationId": "39887e19ff3451cfda2bbf685044107d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserComment"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Post": {
                "description": "Модель поста",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Название поста"
                    },
                    "description": {
                        "type": "string",
                        "example": "Краткое описание"
                    },
                    "text": {
                        "type": "string",
                        "example": "Содержимое поста"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "icon": {
                        "type": "string",
                        "example": "image.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "UserComment": {
                "description": "Модель комментария",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "example": "5"
                    },
                    "title": {
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "text": {
                        "type": "string",
                        "example": "Текст комментария"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}